Extract Valid Grid from Compilation Results

This helper function filters the results from compile_keras_grid() to return a new hyperparameter grid containing only the combinations that compiled successfully.

Usage

extract_valid_grid(compiled_grid)

Arguments

compiled_grid

A tibble, the result of a call to compile_keras_grid().

Value

A tibble containing the subset of the original grid that resulted in a successful model compilation (i.e., where the error column is NA). The columns for compiled_model, model_summary, and error are removed.